+++ daemon-backend/src/app.js
@@ -3,6 +3,8 @@ const cors = require("cors");
 const helmet = require("helmet");
 const compression = require("compression");
 const morgan = require("morgan");
+const projectRoutes = require("./routes/project.routes");
+const errorHandler = require("./middleware/errorHandler");
 
 const app = express();
 
@@ -10,8 +12,10 @@ app.use(cors());
 app.use(helmet());
 app.use(compression());
 app.use(morgan("dev"));
-
 app.use(express.json());
+
+app.use("/projects", projectRoutes);
+
 
 app.get("/", (req, res) => {
   res.json({
@@ -20,5 +24,5 @@ app.get("/", (req, res) => {
     status: "Running 🚀"
   });
 });
-
+app.use(errorHandler);
 module.exports = app;